refactor: revise Home page
