--- pages/home/index.tsx
@@ -1,1 +1,0 @@
-export { default } from '../../components/templates/Home';
